mail: respect `mail.activity` ir.rules when reading progress bar
In order to reproduce this issue, one would need to customize the `ir.rules` of
the `mail.activity` model in order to, for example, hide the activities from
other users.

In such situation, the `_read_progress_bar` optimization in the mixin was
completely ignoring the `ir.rules` of `mail.activity` records.

This commit fixes the raw SQL in order to account for them.