account: Allow different payment accounts per journal
If you assing a different payment account per payment method line
corresponding to the same journal (but different payment methods),

The method that returns the possible liquidity accounts returns a tuple
instead of a recordset, and the comparison is done with the operator
`in`, so if any of the elements of the tuple is a recordset of more
than one record, the match is not happening, which may be the case for
`self.journal_id.inbound_payment_method_line_ids.payment_account_id` or
`self.journal_id.outbound_payment_method_line_ids.payment_account_id`.

The solution is to return a recordset instead.

TT43014

closes odoo/odoo#120847

Signed-off-by: William André (wan) <wan@odoo.com>