product: improve perf of `product.pricelist._is_applicable_for`
This commit introduces the following changes:

* Early exit in case of `False` result.
* Leverage `parent_path` for category-based rules.
* Use `applied_on` for the conditions, which makes the code more readable.

On a database with 470 products, and about the same number of pricelist items,
this results in a 30% speedup when computing prices for all products at once. It
may be even better depending on how many nested product categories are used in
the database and in the pricelist rules.