account/portal: Prioritize user in order of recipient groups
In case a user logs an internal note on an invoice and pings
another internal user, who happens to manage notifications by email,
the link to the document inside the email was directing to the portal
instead of the backend.

This happens because of the override of _notify_get_groups in the
account module adds an extra recipient group 'additional_intended_recipient'
in the first position of the groups list. This group is based on
'portal_customer' but its validation function will evaluate to True for
internal users.

In any case, the recipient group for user must be considered first as
only internal user will be validated through its function. The recipient
group portal_customer can then be added in the second position, ie
after user but before portal, and the eventual additional_intended_recipient
can be set before portal_customer as it is the case now.

Steps to reproduce:
1. Have a internal user that manages notifications by email
2. On a customer invoice log a note and ping this user using '@'

Before:
The email contains a link to the portal

After:
The email contains a link to the backend