account: Avoid resetting balance when amount currency is zero
Without this commit, in case a move line is created with a defined
balance and an amount currency equals 0, an error will be raised due
to the _check_balanced constraint.

This happens, because credit / debit are first computed based on the
defined balance, but the _inverse_amount_currency function will then
reset the balance to 0, what will raise an error due to wrongly
unbalanced moves.